+++ packages/bridge-controller/src/utils/metrics/properties.test.ts
@@ -1,7 +1,8 @@
 import { SolScope } from '@metamask/keyring-api';
 import type { CaipChainId } from '@metamask/utils';
 
-import type { QuoteMetadata, QuoteResponseV1 } from '../../types';
+import type { QuoteMetadata } from '../../types';
+import type { QuoteResponseV1 } from '../../validators/quote-response-v1';
 import { getNativeAssetForChainId } from '../bridge';
 import { formatChainIdToCaip } from '../caip-formatters';
 import { MetricsSwapType } from './constants';
