+++ packages/bridge-controller/src/validators/feature-flags.ts
@@ -7,7 +7,6 @@ import {
   array,
   boolean,
   number,
-  enums,
   is,
   define,
 } from '@metamask/superstruct';
@@ -74,7 +73,6 @@ export const ChainConfigurationSchema = type({
 });
 
 export const PriceImpactThresholdSchema = type({
-  // TODO:
   // We are moving into a unified approach where
   // price impact thresholds will be segmented by
   // importance rather than transaction type.
@@ -89,20 +87,20 @@ export const PriceImpactThresholdSchema = type({
   warning: optional(number()), // Percentage value in decimal format
   error: optional(number()), // Percentage value in decimal format
 });
+
 const GenericQuoteRequestSchema = type({
   aggIds: optional(array(string())),
   bridgeIds: optional(array(string())),
   fee: optional(number()),
 });
-const FeatureIdSchema = enums(Object.values(FeatureId));
+
 /**
  * This is the schema for the feature flags response from the RemoteFeatureFlagController
  */
-
 export const PlatformConfigSchema = type({
   priceImpactThreshold: optional(PriceImpactThresholdSchema),
   quoteRequestOverrides: optional(
-    record(FeatureIdSchema, optional(GenericQuoteRequestSchema)),
+    record(string(), optional(GenericQuoteRequestSchema)),
   ),
   minimumVersion: string(),
   refreshRate: number(),
