+++ packages/bridge-controller/src/validators/validators.test.ts
@@ -2,12 +2,10 @@ import { is } from '@metamask/superstruct';
 
 import { mockBridgeQuotesNativeErc20EthV1 } from '../../tests/mock-quotes-native-erc20-eth';
 import { validateFeatureFlagsResponse } from './feature-flags';
-import {
-  IntentSchema,
-  validateQuoteResponseV1,
-  FeeDataSchema,
-  DiscountType,
-} from './quote-response';
+import { IntentSchema } from './intent';
+import { DiscountType } from './quote';
+import { FeeDataSchema } from './quote';
+import { validateQuoteResponseV1 } from './quote-response-v1';
 import {
   validateQuoteStreamComplete,
   QuoteStreamCompleteReason,
